feat(conformance,control-plane): extend protocol harness, audience token, and conformance suite
- conformance/fsqlite_inspired_suite.rs: extend conformance suite
- conformance/protocol_harness.rs: expand protocol harness logic
- control_plane/audience_token.rs: extend audience token handling (+24/-2)
- control_plane/transition_abort.rs: minor refinement
- .beads/issues.jsonl: sync tracker state

+++ crates/franken-node/src/control_plane/transition_abort.rs
@@ -210,7 +210,7 @@ impl ForceTransitionPolicy {
             sha2::Digest::update(&mut hasher, (field.len() as u64).to_le_bytes());
             sha2::Digest::update(&mut hasher, field.as_bytes());
         }
-        sha2::Digest::update(&mut hasher, self.max_skippable.to_le_bytes());
+        sha2::Digest::update(&mut hasher, (self.max_skippable as u64).to_le_bytes());
         format!("policy:{:x}", sha2::Digest::finalize(hasher))
     }
 }
